# api/api.yaml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: TorrPlay API
  description: A torrent streaming application with memory-managed piece storage.
  termsOfService: https://example.com/terms/
  contact:
    name: TorrPlay Team
    email: support@example.com
    url: https://github.com/torrplay
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
security:
  - basicAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v2/torrents/add:
    post:
      tags:
        - qBittorrent
      summary: Add a new torrent (qBittorrent compatible).
      description: Compatibility endpoint for adding a new torrent.
      operationId: qBittorrentAddTorrent
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                category:
                  type: string
                  description: Category for the torrent.
                torrents:
                  type: array
                  items:
                    type: string
                    format: binary
                urls:
                  type: string
                  description: A list of magnet links to add.
      responses:
        "200":
          description: Torrent added successfully.
          content:
            text/plain:
              schema:
                type: string
                example: "Ok."
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"

  # TorrServer Compatibility Endpoints
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
  responses:
...
    BadRequest:
      description: The request was invalid or cannot be served.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid info hash format."
            code: 400
            details: "Info hash must be 40 hex characters."
    Unauthorized:
      description: Authentication failed or is missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized."
            code: 401
            details: "Authentication token is missing or invalid."
...
    InternalServerError:
      description: An internal server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error."
            code: 500
            details: "Failed to allocate memory for piece storage."
...
    GatewayTimeout:
      description: The server did not receive a timely response from an upstream server.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Gateway timeout."
            code: 504
            details: "Timeout waiting for torrent metadata."
  schemas:
...
    Error:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
          description: Human-readable error message.
          example: "Torrent not found"
        code:
          type: integer
          description: HTTP status code.
          minimum: 400
          maximum: 599
          example: 404
...